Initial HTTP proxy handling, normal and transparent
Introduce an early version of a proxy that actually speaks HTTP
internally, and is not just a dumb TCP proxy. This handles CONNECT
requests as well as GET requests with a path and a host header (for
transparent proxying) and of course a GET request with a URL in the
path component. HEAD methods are supported too, but it is unclear which
other methods are working too.

The implementation is a bit weird: We tell net/http to use the last peer
in the list of peers as a proxy and provide a custom Dial function
to the transport that can only dial to that proxy. This enables reuse
of proxy connections, reducing the overhead of TLS handshakes.

CONNECT is handled specially: We first dial to the upstream proxy and
then hijack the connection and forward bytes between them like the dumb
TCP forwarder does. This ensures that CONNECT works as intended.

Compared to the TCP proxy, the HTTP proxy handler has these advantages:

1. It can handle transparent requests
2. It can automatically inject additional headers into the final proxy
   connection, thus removing the need for e.g. Authorization at the
   client.
3. It can cache and reuse connections to the upstream proxy, thus getting rid of
  the TLS handshake overhead
that it can automatically provide